"Replace default App component with a new component that rounds grades and displays them"
This commit removes the default App component that displays a logo and a link to the React website. It adds a new App component that takes an array of grades, rounds them according to the rules specified in the `roundGrades` function, and displays the rounded grades in an unordered list.